refactor: init folder structure.
