+++ js/form-img-upload/validate-hashtags.js
