+++ js/gallery/big-picture.js
@@ -1,4 +1,4 @@
-import { isEscapeKey } from './util.js';
+import { isEscapeKey } from '../util.js';
 import {
   insertComments,
   showMoreComments,
